# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RunningAI API",
    "version": "1.2.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/model/info": {
      "get": {
        "summary": "Model Info",
        "operationId": "model_info_api_v1_model_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/predict/overstride": {
      "post": {
        "summary": "Predict Overstride",
        "operationId": "predict_overstride_api_v1_predict_overstride_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverstrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverstrideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/predict/strike-type": {
      "post": {
        "summary": "Predict Strike",
        "operationId": "predict_strike_api_v1_predict_strike_type_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrikeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrikeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/analyze": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_api_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/adapter/debug-export": {
      "post": {
        "summary": "Adapt Debug Export",
        "operationId": "adapt_debug_export_api_v1_adapter_debug_export_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_adapt_debug_export_api_v1_adapter_debug_export_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/analyze/debug-export": {
      "post": {
        "summary": "Analyze Debug Export",
        "operationId": "analyze_debug_export_api_v1_analyze_debug_export_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_debug_export_api_v1_analyze_debug_export_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "minLength": 1,
            "title": "Session Id"
          },
          "patient_meta": {
            "$ref": "#/components/schemas/PatientMeta"
          },
          "overstride_features": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Overstride Features"
          },
          "strike_feet": {
            "items": {
              "$ref": "#/components/schemas/StrikeFootInput"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Strike Feet"
          },
          "posture_metrics": {
            "$ref": "#/components/schemas/PostureMetrics"
          },
          "quality": {
            "$ref": "#/components/schemas/QualityInfo"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "patient_meta",
          "overstride_features",
          "strike_feet"
        ],
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "patient_id": {
            "type": "string",
            "minLength": 1,
            "title": "Patient Id"
          },
          "overstride": {
            "$ref": "#/components/schemas/OverstrideResponse"
          },
          "strike_type": {
            "$ref": "#/components/schemas/StrikeResponse"
          },
          "posture_metrics": {
            "$ref": "#/components/schemas/PostureMetrics"
          },
          "quality": {
            "$ref": "#/components/schemas/QualityInfo"
          },
          "model_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Model Info"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "patient_id",
          "overstride",
          "strike_type",
          "posture_metrics",
          "quality",
          "model_info"
        ],
        "title": "AnalyzeResponse"
      },
      "Body_adapt_debug_export_api_v1_adapter_debug_export_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_adapt_debug_export_api_v1_adapter_debug_export_post"
      },
      "Body_analyze_debug_export_api_v1_analyze_debug_export_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_debug_export_api_v1_analyze_debug_export_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelInfoResponse": {
        "properties": {
          "api_version": {
            "type": "string",
            "title": "Api Version"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "supported_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Metrics"
          },
          "final_independent_test_completed": {
            "type": "boolean",
            "title": "Final Independent Test Completed"
          },
          "final_independent_targets_met": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Independent Targets Met"
          },
          "artifact_integrity": {
            "type": "boolean",
            "title": "Artifact Integrity"
          },
          "strike_aggregation": {
            "type": "string",
            "title": "Strike Aggregation"
          },
          "strike_rescue_rule": {
            "type": "string",
            "title": "Strike Rescue Rule"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "api_version",
          "model_version",
          "stage",
          "supported_metrics",
          "final_independent_test_completed",
          "final_independent_targets_met",
          "artifact_integrity",
          "strike_aggregation",
          "strike_rescue_rule",
          "notes"
        ],
        "title": "ModelInfoResponse"
      },
      "OverstrideRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "minLength": 1,
            "title": "Session Id"
          },
          "patient_meta": {
            "$ref": "#/components/schemas/PatientMeta"
          },
          "features": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Features"
          },
          "quality_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "high",
                  "medium",
                  "low"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Tier"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "patient_meta",
          "features"
        ],
        "title": "OverstrideRequest"
      },
      "OverstrideResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "patient_id": {
            "type": "string",
            "minLength": 1,
            "title": "Patient Id"
          },
          "prediction_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Mm"
          },
          "prediction_std_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Std Mm"
          },
          "model_pair_count": {
            "type": "integer",
            "title": "Model Pair Count"
          },
          "quality_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "high",
                  "medium",
                  "low"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Tier"
          },
          "review_required": {
            "type": "boolean",
            "title": "Review Required"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "unmeasurable"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "patient_id",
          "prediction_mm",
          "prediction_std_mm",
          "model_pair_count",
          "quality_tier",
          "review_required",
          "status",
          "model_version"
        ],
        "title": "OverstrideResponse"
      },
      "PatientMeta": {
        "properties": {
          "patient_id": {
            "type": "string",
            "minLength": 1,
            "title": "Patient Id"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "running_speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Running Speed Kmh"
          },
          "video_fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Fps"
          }
        },
        "type": "object",
        "required": [
          "patient_id"
        ],
        "title": "PatientMeta"
      },
      "PostureMetrics": {
        "properties": {
          "cadence_spm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cadence Spm"
          },
          "contact_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Time Ms"
          },
          "forward_lean_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Lean Deg"
          },
          "max_thigh_flexion_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Thigh Flexion Deg"
          },
          "max_thigh_extension_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Thigh Extension Deg"
          },
          "knee_flexion_touchdown_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knee Flexion Touchdown Deg"
          },
          "pelvic_drop_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pelvic Drop Deg"
          },
          "hip_hike_difference_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hip Hike Difference Deg"
          },
          "shank_angle_touchdown_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shank Angle Touchdown Deg"
          }
        },
        "type": "object",
        "title": "PostureMetrics"
      },
      "QualityInfo": {
        "properties": {
          "quality_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "high",
                  "medium",
                  "low"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Tier"
          },
          "side_pose_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Side Pose Rate"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "QualityInfo"
      },
      "StrikeEvent": {
        "properties": {
          "sequence": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array",
              "maxItems": 17,
              "minItems": 17
            },
            "type": "array",
            "maxItems": 21,
            "minItems": 21,
            "title": "Sequence"
          },
          "rule_class": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "heel",
                  "midfoot",
                  "forefoot"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Class"
          }
        },
        "type": "object",
        "required": [
          "sequence"
        ],
        "title": "StrikeEvent"
      },
      "StrikeFootInput": {
        "properties": {
          "foot": {
            "type": "string",
            "enum": [
              "left",
              "right"
            ],
            "title": "Foot"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/StrikeEvent"
            },
            "type": "array",
            "minItems": 1,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "foot",
          "events"
        ],
        "title": "StrikeFootInput"
      },
      "StrikeFootResult": {
        "properties": {
          "foot": {
            "type": "string",
            "enum": [
              "left",
              "right"
            ],
            "title": "Foot"
          },
          "prediction": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "heel",
                  "midfoot",
                  "forefoot"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "review_required": {
            "type": "boolean",
            "title": "Review Required"
          },
          "local_probabilities": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Probabilities"
          }
        },
        "type": "object",
        "required": [
          "foot",
          "prediction",
          "confidence",
          "review_required"
        ],
        "title": "StrikeFootResult"
      },
      "StrikeRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "minLength": 1,
            "title": "Session Id"
          },
          "patient_meta": {
            "$ref": "#/components/schemas/PatientMeta"
          },
          "feet": {
            "items": {
              "$ref": "#/components/schemas/StrikeFootInput"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Feet"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "patient_meta",
          "feet"
        ],
        "title": "StrikeRequest"
      },
      "StrikeResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "patient_id": {
            "type": "string",
            "minLength": 1,
            "title": "Patient Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "unmeasurable"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "feet": {
            "items": {
              "$ref": "#/components/schemas/StrikeFootResult"
            },
            "type": "array",
            "title": "Feet"
          },
          "patient_anchor_class": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "heel",
                  "midfoot",
                  "forefoot"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Anchor Class"
          },
          "patient_anchor_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Anchor Confidence"
          },
          "final_class": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "heel",
                  "midfoot",
                  "forefoot"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Class"
          },
          "final_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Confidence"
          },
          "review_required": {
            "type": "boolean",
            "title": "Review Required"
          },
          "rescue_p_heel": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rescue P Heel"
          },
          "rescue_activated": {
            "type": "boolean",
            "title": "Rescue Activated",
            "default": false
          },
          "aggregation": {
            "type": "string",
            "title": "Aggregation",
            "default": "patient_consensus"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "patient_id",
          "status",
          "feet",
          "review_required",
          "model_version"
        ],
        "title": "StrikeResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "RunningAI API token",
        "scheme": "bearer"
      }
    }
  }
}